Change entity API
